requirements.txt
Django==2.0.3
django-compat==1.0.15
django-extensions==2.0.5
django-hijack==2.1.7
django-material==1.2.2
Pillow==5.0.0
psycopg2==2.7.3
pytz==2018.3
six==1.11.0
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
